+++ server/Dockerfile
@@ -1,7 +1,14 @@
 # syntax=docker/dockerfile:1
 FROM python:3.9-slim-buster
 
+# Install Tesseract-OCR
+#If you are using a different Linux distribution, you’ll need to copy the last github repository version and copy the
+#.traineddata file into ‘tessdata’ (/usr/share/tesseract-ocr/tessdata or /usr/share/tessdata).
+RUN apt-get install tesseract-ocr
+RUN apt-get install imagemagick
+
 WORKDIR /server
+
 COPY requirements.txt requirements.txt
 RUN pip3 install -r requirements.txt
 COPY . .
